+++ RELEASES.md
@@ -2,6 +2,7 @@ Version 0.8.11 (2021-XX-XX)
 ========================
  * Navigation: Show more detailed instructions when approaching maneuver
  * Cadillac Escalade ESV 2016 support thanks to Gibby!
+ * Mazda CX-9 2021 support thanks to Jacar!
 
 Version 0.8.10 (2021-11-01)
 ========================